Trivia Question Database: almost there

+++ app/src/main/java/algonquin/cst2335/finalproject/QuizContestant.java
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey;
 public class QuizContestant {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id")
-    public long id;
+    public int id;
     @ColumnInfo(name = "score")
     public String score;
     @ColumnInfo(name = "name")
@@ -16,17 +16,17 @@ public class QuizContestant {
 
     public QuizContestant() {}
 
-    public QuizContestant(long id, String score, String name) {
+    public QuizContestant(int id, String score, String name) {
         this.id = id;
         this.score = score;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
